requirements.txt
Django
djangorestframework
drf-spectacular
django-cors-headers
psycopg2-binary
gunicorn
whitenoise